Revert Commit 'Add FBX SDK' from 'FbxTest make fails on unit test' (35af47c..52981fa). Importing fbx sdk from scratch

+++ HopStep/HopStepEngine/Runtime/Engine/EngineLoop.cpp
@@ -23,7 +23,7 @@ namespace HopStep
 
         // Engine Tick
         GEngine->UpdateTime();
-        GEngine->Tick(static_cast<float>(HApp::GetDeltaTime()));
+        GEngine->Tick(HApp::GetDeltaTime());
     }
 
     void HEngineLoop::Exit()
